\name{pmg-package}
\alias{pmg-package}
\docType{package}
\title{
Poor Man's GUI
}
\description{
Simple GUI for R using RGtk2 and iWidgetsRGtk
}
\details{
  
Further information is available in the following vignettes:
\tabular{ll}{
\code{manual} \tab pmg (source, pdf)\cr
}
}
\author{
John Verzani
Maintainer: John Verzani <pmgrgui@gmail.com>
URL: \url{http://www.math.csi.cuny.edu/pmg}
}
\keyword{ package }
